Remove seemingly unnecessary placeholder
This placeholder presence causes the drag of the last item to display an
empty <li> element. That empty <li> element renders the bullet point so
the dragging appears broken.

If the item is being dragged to not be the last item but somwehere else
then this bug is not visible.